Charge and barrage attack ported

+++ src/main/java/insane96mcp/progressivebosses/module/wither/data/WitherAttackStats.java
@@ -14,7 +14,7 @@ public class WitherAttackStats {
     public int attackSpeedFar;
     public float maxChargeChance;
     public float chargeDamage;
-    public float chargeTime;
+    public int chargeTime;
     public float barrageChance;
     public int barrageDuration;
